use gray-matter to show blogs page
