+++ data-sources/blogs/hello-world.md
@@ -1,3 +1,11 @@
+---
+id: 'hello-world'
+title: 'ブログ作りました'
+description: 'ブログ作りました' 
+createdAt: '2019-10-04'
+updatedAt: '2020-04-21'
+---
+
 ## hello world
 
 ```javascript
