+++ lib/blog-data-list.ts
@@ -1,11 +1,4 @@
-export interface BlogData {
-  id: string;
-  title: string;
-  description: string;
-  createdAt: string;
-  updatedAt: string;
-  image?: string;
-}
+import { BlogData } from '../src/models';
 
 export const blogDataList: BlogData[] = [
   { id: 'hello-world', title: 'ブログ作りました', description: 'ブログ作りました', createdAt: '2019-10-04', updatedAt: '2020-04-21' },
