+++ src/models.ts
@@ -1,0 +1,8 @@
+export interface BlogData {
+  id: string;
+  title: string;
+  description: string;
+  createdAt: string;
+  updatedAt: string;
+  image?: string;
+}
